+++ src/effects/transforms/compound.ts
@@ -4,6 +4,7 @@ import * as col from "./color";
 import * as con from "./content";
 import * as rot from "./rotations";
 import { CellTransformationFunction } from "./index";
+import { moveEverythingViaTranslate } from "./move";
 
 type F = CellTransformationFunction;
 
@@ -25,4 +26,7 @@ export default [
     rot.transformSingleRotateAll,
     col.transformNRandomColors
   ),
+  compound(moveEverythingViaTranslate, rot.transformSingleRotateZ),
+  compound(moveEverythingViaTranslate, rot.transformSingleRotateAll),
+  compound(moveEverythingViaTranslate, rot.transformIndependentRotateAll),
 ] as const;
